feat(#1078): ICA: ripgrep-search args + parse single-consumer pure-function modules
